add authorization checks when updating and deleting a post record; assign the authorized user's id to the post's owne r_id when creating a post

+++ app/router/post.py
@@ -5,6 +5,8 @@ from ..database import get_db
  
 router = APIRouter(prefix="/posts", tags=['Posts'])
 
+post_authorization_exception = HTTPException(status.HTTP_403_FORBIDDEN, "you do not own this post")
+
 @router.get("/", response_model=list[schemas.PostResponse])
 def get_posts(db: Session = Depends(get_db)):
     posts = db.query(models.Post).all()
@@ -16,12 +18,15 @@ def get_post(id: int, db: Session = Depends(get_db)):
 
     if not post:
         raise HTTPException(status.HTTP_404_NOT_FOUND, f"post with id '{id}' does not exist")
-    return {"data": post}
+    return post
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.PostResponse)
 def create_post(payload: schemas.CreatePost, db: Session = Depends(get_db), req_user: models.User = Depends(oauth2.get_current_user)):
     # unpack a dict: **dict
-    post = models.Post(**payload.model_dump())
+    post_data = payload.model_dump()
+    post_data.update({"owner_id": req_user.id})
+
+    post = models.Post(**post_data)
     db.add(post)
     db.commit()
     db.refresh(post)
@@ -35,13 +40,15 @@ def put_post(id: int, payload: schemas.CreatePost, db: Session = Depends(get_db)
 
     if not post:
         raise HTTPException(status.HTTP_404_NOT_FOUND, f"post with id '{id}' does not exist")
+    if post.owner_id != req_user.id:
+        raise post_authorization_exception
     
     post_query.update(payload.model_dump(), synchronize_session=False)
     db.commit()
 
     return post_query.first()
 
-@router.patch("/{id}")
+@router.patch("/{id}", response_model=schemas.PostResponse)
 def patch_post(id: int, payload: schemas.UpdatePost, db: Session = Depends(get_db), req_user: models.User = Depends(oauth2.get_current_user)):
     # filter out None values, dictionary comprehension
     filtered_data = { k: v for k, v in payload.model_dump().items() if v is not None }
@@ -50,6 +57,8 @@ def patch_post(id: int, payload: schemas.UpdatePost, db: Session = Depends(get_d
     post = post_query.first()
     if not post:
         raise HTTPException(status.HTTP_404_NOT_FOUND, f"post with id '{id}' does not exist")
+    if post.owner_id != req_user.id:
+        raise post_authorization_exception
     
     post_query.update(filtered_data, synchronize_session=False)
     db.commit()
@@ -60,8 +69,11 @@ def patch_post(id: int, payload: schemas.UpdatePost, db: Session = Depends(get_d
 def delete_post(id: int, db: Session = Depends(get_db), req_user: models.User = Depends(oauth2.get_current_user)):
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
+
     if not post:
         raise HTTPException(status.HTTP_404_NOT_FOUND, f"post with id '{id}' does not exist")
+    if post.owner_id != req_user.id:
+        raise post_authorization_exception
     
     post_query.delete(synchronize_session=False)
     db.commit()
